--- src/main/java/org/cheonyakplanet/be/domain/entity/UserStatusEnum.java
@@ -1,7 +1,0 @@
-package org.cheonyakplanet.be.domain.entity;
-
-public enum UserStatusEnum {
-
-  ACTIVE,  // 기본 상태
-  WITHDRAWN  // 탈퇴 상태
-}
