+++ src/main/java/org/cheonyakplanet/be/domain/entity/comunity/Post.java
@@ -1,0 +1,81 @@
+package org.cheonyakplanet.be.domain.entity.comunity;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import org.cheonyakplanet.be.application.dto.community.PostDTO;
+import org.cheonyakplanet.be.domain.Stamped;
+
+import com.fasterxml.jackson.annotation.JsonManagedReference;
+
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Table(catalog = "planet", name = "Post")
+public class Post extends Stamped {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+
+	private String username;
+
+	private String title;
+
+	@Column(columnDefinition = "TEXT")
+	private String content;
+
+	private Long views; // 조회수
+
+	private int likes; // 추천 수
+	private int dislikes; //싫어요 수
+
+	@Builder.Default
+	@JsonManagedReference
+	@OneToMany(mappedBy = "post", cascade = CascadeType.ALL, orphanRemoval = true)
+	private List<Comment> comments = new ArrayList<>();
+
+	@Builder.Default
+	private boolean isBlind = false;
+
+	public void incrementLikes() {
+		this.likes++;
+	}
+
+	public void increaseDislikes() {
+		this.dislikes--;
+	}
+
+	public void countViews() {
+		this.views = (this.views == null ? 1L : this.views + 1);
+	}
+
+	public static PostDTO ToDTO(Post post) {
+		return PostDTO.builder()
+			.id(post.getId())
+			.username(post.getUsername())
+			.title(post.getTitle())
+			.content(post.getContent())
+			.views(post.getViews())
+			.likes(post.getLikes())
+			.createdAt(post.getCreatedAt())
+			.build();
+	}
+
+}
