+++ src/main/java/org/cheonyakplanet/be/domain/entity/comunity/PostReaction.java
@@ -1,0 +1,39 @@
+package org.cheonyakplanet.be.domain.entity.comunity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Table(name = "post_reaction", catalog = "planet")
+public class PostReaction {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+
+	@Column(nullable = false)
+	private String email;  // 반응을 남긴 사용자의 이메일
+
+	@ManyToOne
+	@JoinColumn(name = "post_id", nullable = false)
+	private Post post;
+
+	@Enumerated(EnumType.STRING)
+	@Column(nullable = false)
+	private ReactionType reactionType;
+}
