+++ src/main/java/org/cheonyakplanet/be/domain/entity/comunity/Reply.java
@@ -1,0 +1,40 @@
+package org.cheonyakplanet.be.domain.entity.comunity;
+
+import org.cheonyakplanet.be.domain.Stamped;
+
+import com.fasterxml.jackson.annotation.JsonBackReference;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Table(catalog = "planet", name = "Reply")
+public class Reply extends Stamped {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+
+	@Column(columnDefinition = "TEXT")
+	private String content;
+
+	@JsonBackReference
+	@ManyToOne
+	@JoinColumn(name = "comment_id")
+	private Comment comment;
+}
