+++ src/main/java/org/cheonyakplanet/be/domain/entity/user/UserStatusEnum.java
@@ -1,0 +1,7 @@
+package org.cheonyakplanet.be.domain.entity.user;
+
+public enum UserStatusEnum {
+
+	ACTIVE,  // 기본 상태
+	WITHDRAWN  // 탈퇴 상태
+}
