+++ src/main/java/org/cheonyakplanet/be/domain/repository/CommentRepository.java
@@ -1,6 +1,6 @@
 package org.cheonyakplanet.be.domain.repository;
 
-import org.cheonyakplanet.be.domain.entity.Comment;
+import org.cheonyakplanet.be.domain.entity.comunity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
