+++ src/main/java/org/cheonyakplanet/be/domain/repository/ReplyRepository.java
@@ -1,6 +1,6 @@
 package org.cheonyakplanet.be.domain.repository;
 
-import org.cheonyakplanet.be.domain.entity.Reply;
+import org.cheonyakplanet.be.domain.entity.comunity.Reply;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
