+++ src/main/java/org/cheonyakplanet/be/domain/repository/UserTokenRepository.java
@@ -1,19 +1,19 @@
 package org.cheonyakplanet.be.domain.repository;
 
-import org.cheonyakplanet.be.domain.entity.UserToken;
+import java.util.Optional;
+
+import org.cheonyakplanet.be.domain.entity.user.UserToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
 
 @Repository
 public interface UserTokenRepository extends JpaRepository<UserToken, Long> {
 
-    Optional<UserToken> findByEmail(String token);
+	Optional<UserToken> findByEmail(String token);
 
-    Optional<UserToken> findByRefreshToken(String token);
+	Optional<UserToken> findByRefreshToken(String token);
 
-    Optional<UserToken> findByAccessToken(String accessToken);
+	Optional<UserToken> findByAccessToken(String accessToken);
 
-    void deleteByEmail(String email);
+	void deleteByEmail(String email);
 }
