+++ src/main/java/org/cheonyakplanet/be/infrastructure/jwt/JwtAuthenticationFilter.java
@@ -1,90 +1,94 @@
 package org.cheonyakplanet.be.infrastructure.jwt;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+import java.io.IOException;
+
 import org.cheonyakplanet.be.application.dto.user.LoginRequestDTO;
-import org.cheonyakplanet.be.domain.entity.UserRoleEnum;
+import org.cheonyakplanet.be.domain.entity.user.UserRoleEnum;
 import org.cheonyakplanet.be.infrastructure.security.UserDetailsImpl;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
-import java.io.IOException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
 @Slf4j(topic = "로그인 및 JWT 생성")
 @RequiredArgsConstructor
 public class JwtAuthenticationFilter extends UsernamePasswordAuthenticationFilter {
 
-    private final JwtUtil jwtUtil;
+	private final JwtUtil jwtUtil;
 
-    /**
-     * 로그인 필터
-     * @param request
-     * @param response
-     * @return
-     * @throws AuthenticationException
-     */
-    @Override
-    public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) throws AuthenticationException {
-        log.info("로그인 시도: Content-Type = {}", request.getContentType());
-        try {
-            LoginRequestDTO requestDto = new ObjectMapper().readValue(request.getInputStream(), LoginRequestDTO.class);
+	/**
+	 * 로그인 필터
+	 * @param request
+	 * @param response
+	 * @return
+	 * @throws AuthenticationException
+	 */
+	@Override
+	public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) throws
+		AuthenticationException {
+		log.info("로그인 시도: Content-Type = {}", request.getContentType());
+		try {
+			LoginRequestDTO requestDto = new ObjectMapper().readValue(request.getInputStream(), LoginRequestDTO.class);
 
-            log.info("로그인 요청 데이터: {}", requestDto);
+			log.info("로그인 요청 데이터: {}", requestDto);
 
-            return getAuthenticationManager().authenticate(
-                    new UsernamePasswordAuthenticationToken(
-                            requestDto.getEmail(),
-                            requestDto.getPassword(),
-                            null
-                    )
-            );
-        } catch (IOException e) {
-            log.error("요청 데이터 읽기 실패", e);
-            throw new RuntimeException(e.getMessage());
-        }
-    }
+			return getAuthenticationManager().authenticate(
+				new UsernamePasswordAuthenticationToken(
+					requestDto.getEmail(),
+					requestDto.getPassword(),
+					null
+				)
+			);
+		} catch (IOException e) {
+			log.error("요청 데이터 읽기 실패", e);
+			throw new RuntimeException(e.getMessage());
+		}
+	}
 
-    /**
-     * 로그인 성공시
-     * @param request
-     * @param response
-     * @param chain
-     * @param authResult
-     * @throws IOException
-     * @throws ServletException
-     */
-    @Override
-    protected void successfulAuthentication(HttpServletRequest request, HttpServletResponse response, FilterChain chain, Authentication authResult) throws IOException, ServletException {
-        log.info("로그인 성공 및 JWT 생성");
-        String email = ((UserDetailsImpl) authResult.getPrincipal()).getUser().getEmail();
-        UserRoleEnum role = ((UserDetailsImpl) authResult.getPrincipal()).getUser().getRole();
+	/**
+	 * 로그인 성공시
+	 * @param request
+	 * @param response
+	 * @param chain
+	 * @param authResult
+	 * @throws IOException
+	 * @throws ServletException
+	 */
+	@Override
+	protected void successfulAuthentication(HttpServletRequest request, HttpServletResponse response, FilterChain chain,
+		Authentication authResult) throws IOException, ServletException {
+		log.info("로그인 성공 및 JWT 생성");
+		String email = ((UserDetailsImpl)authResult.getPrincipal()).getUser().getEmail();
+		UserRoleEnum role = ((UserDetailsImpl)authResult.getPrincipal()).getUser().getRole();
 
-        String token = jwtUtil.createAccessToken(email, role);
-        response.addHeader(JwtUtil.AUTHORIZATION_HEADER, token);
+		String token = jwtUtil.createAccessToken(email, role);
+		response.addHeader(JwtUtil.AUTHORIZATION_HEADER, token);
 
-        response.sendRedirect("/");
-    }
+		response.sendRedirect("/");
+	}
 
-    /**
-     * 로그인 실패시
-     * @param request
-     * @param response
-     * @param failed
-     * @throws IOException
-     * @throws ServletException
-     */
-    @Override
-    protected void unsuccessfulAuthentication(HttpServletRequest request, HttpServletResponse response, AuthenticationException failed) throws IOException, ServletException {
-        log.info("로그인 실패");
-        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-    }
-
+	/**
+	 * 로그인 실패시
+	 * @param request
+	 * @param response
+	 * @param failed
+	 * @throws IOException
+	 * @throws ServletException
+	 */
+	@Override
+	protected void unsuccessfulAuthentication(HttpServletRequest request, HttpServletResponse response,
+		AuthenticationException failed) throws IOException, ServletException {
+		log.info("로그인 실패");
+		response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+	}
 
 }
